requirements.txt
python-telegram-bot==22.6
Pillow==12.1.0
img2pdf==0.6.3
python-docx==0.8.11
PyPDF2==3.0.1
reportlab==4.3.0